PCA: class _LinearCombination moved to base
